Fixed request parameters for 'arc'
