+++ src/requests.rs
@@ -1,5 +1,7 @@
+mod info;
+mod parameter;
 mod request;
-mod request_info;
 
+pub use info::*;
+pub use parameter::*;
 pub use request::*;
-pub use request_info::*;
